+++ doc/NEWS.md
@@ -1,6 +1,12 @@
 Scroll View Release Notes
 ==========================
 
+Version 1.0.2 - Jun 16, 2015
+---------------------
+
+- Fix pointer leave event, allowing to scroll content after pointer was away.
+- Add support for smart scrollbar that hides when content fits into element.
+
 Version 1.0.1 - Jun 11, 2015
 ---------------------